fix(NodeInfo): chain_id type Uint64

+++ crates/jsonrpc-types/src/godwoken.rs
@@ -1267,7 +1267,7 @@ pub struct NodeRollupConfig {
     pub challenge_maturity_blocks: Uint64,
     pub finality_blocks: Uint64,
     pub reward_burn_rate: Uint32,
-    pub chain_id: Uint32,
+    pub chain_id: Uint64,
 }
 
 #[derive(Clone, Serialize, Deserialize, PartialEq, Eq, Hash, Debug, Default)]
